refactor: pieces 폴더에 필요없는 class 및 enum 추출
