+++ src/main/java/chess/constant/Color.java
@@ -1,4 +1,4 @@
-package chess.pieces;
+package chess.constant;
 
 public enum Color {
     WHITE, BLACK, NO_COLOR
